# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface CatalogItemInput {
  id?: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  emoji: string;
  /** @minimum 1 */
  defaultQty: number;
  /** @minLength 1 */
  unit: string;
  note?: string;
}

export interface ListItemInput {
  itemId: string;
  /** @minimum 1 */
  quantity?: number;
}

export interface HouseholdCreateInput {
  catalog?: CatalogItemInput[];
  list?: ListItemInput[];
}

export interface HouseholdJoinInput {
  /** @pattern ^CASA-[A-Z0-9]{4}$ */
  code: string;
}

export interface ListItemUpdate {
  /** @minimum 1 */
  quantity?: number;
  bought?: boolean;
}

export interface ReceiptInput {
  qrUrl: string;
}

export type CatalogItem = CatalogItemInput & {
  id: string;
  updatedAt: string;
};

export interface ListItem {
  id: string;
  itemId: string;
  quantity: number;
  bought: boolean;
  /** @nullable */
  boughtAt?: string | null;
  item?: CatalogItem;
}

export interface HistoryRecord {
  id: number;
  itemName: string;
  emoji: string;
  purchasedAt: string;
  /** @nullable */
  priceCents?: number | null;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  supermarket?: string | null;
  source: string;
  /** @nullable */
  receiptId?: string | null;
}

export interface Receipt {
  id: string;
  qrUrl: string;
  /** @nullable */
  supermarket?: string | null;
  /** @nullable */
  issuedAt?: string | null;
  /** @nullable */
  totalCents?: number | null;
  importedAt: string;
  resolved: boolean;
}

export interface HouseholdState {
  code: string;
  catalog: CatalogItem[];
  list: ListItem[];
  history: HistoryRecord[];
  receipts: Receipt[];
}

export interface ReceiptResult {
  receipt: Receipt;
  message: string;
  state: HouseholdState;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

export type HouseHeaderParameter = string;