# lib/api-zod/src/generated/types/catalogItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItemInput } from './catalogItemInput';

export type CatalogItem = CatalogItemInput & {
  id: string;
  updatedAt: string;
};