# lib/api-zod/src/generated/types/catalogItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogItemInput {
  id?: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  emoji: string;
  /** @minimum 1 */
  defaultQty: number;
  /** @minLength 1 */
  unit: string;
  note?: string;
}