# lib/api-zod/src/generated/types/historyRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryRecord {
  id: number;
  itemName: string;
  emoji: string;
  purchasedAt: string;
  /** @nullable */
  priceCents?: number | null;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  supermarket?: string | null;
  source: string;
  /** @nullable */
  receiptId?: string | null;
}