# lib/api-zod/src/generated/types/householdCreateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItemInput } from './catalogItemInput';
import type { ListItemInput } from './listItemInput';

export interface HouseholdCreateInput {
  catalog?: CatalogItemInput[];
  list?: ListItemInput[];
}