# lib/api-zod/src/generated/types/householdJoinInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */

export interface HouseholdJoinInput {
  /** @pattern ^CASA-[A-Z0-9]{4}$ */
  code: string;
}