# lib/api-zod/src/generated/types/listItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItem } from './catalogItem';

export interface ListItem {
  id: string;
  itemId: string;
  quantity: number;
  bought: boolean;
  /** @nullable */
  boughtAt?: string | null;
  item?: CatalogItem;
}