# lib/api-zod/src/generated/types/listItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Compra em Casa
 * OpenAPI spec version: 0.1.0
 */

export interface ListItemInput {
  itemId: string;
  /** @minimum 1 */
  quantity?: number;
}